+++ src/Phan/Language/Element/Comment.php
@@ -7,6 +7,7 @@ use Phan\Config;
 use Phan\Language\Context;
 use Phan\Language\Element\Comment\Builder;
 use Phan\Language\Element\Comment\Parameter as CommentParameter;
+use Phan\Language\Element\Comment\Property as CommentProperty;
 use Phan\Language\Element\Comment\Method as CommentMethod;
 use Phan\Language\Element\Flags;
 use Phan\Language\Type;
@@ -24,6 +25,8 @@ use AssertionError;
  * TODO: Pass the doccomment line's index to the Element that will use the client,
  * so that it can be used for more precise line numbers (E.g. for where magic methods were declared,
  * where functions with no signature types but phpdoc types declared types that are invalid class names, etc.
+ *
+ * @see Builder for the logic to create an instance of this class.
  */
 class Comment
 {
@@ -117,7 +120,7 @@ class Comment
     private $suppress_issue_list = [];
 
     /**
-     * @var array<string,CommentParameter>
+     * @var array<string,CommentProperty>
      * A mapping from magic property parameters to types.
      */
     private $magic_property_map = [];
@@ -166,7 +169,7 @@ class Comment
      * @param array<int,string> $suppress_issue_list
      * A list of tags for error type to be suppressed
      *
-     * @param array<int,CommentParameter> $magic_property_list
+     * @param array<int,CommentProperty> $magic_property_list
      *
      * @param array<int,CommentMethod> $magic_method_list
      *
@@ -536,19 +539,19 @@ class Comment
     }
 
     /**
-     * @return CommentParameter
-     * The magic property with the given name. May or may not have a type.
+     * Returns the magic property with the given name.
+     * May or may not have a type.
      * @unused
      * @suppress PhanUnreferencedPublicMethod not used right now, but making it available for plugins
      */
     public function getMagicPropertyWithName(
         string $name
-    ) : CommentParameter {
+    ) : CommentProperty {
         return $this->magic_property_map[$name];
     }
 
     /**
-     * @return array<string,CommentParameter> map from parameter name to parameter
+     * @return array<string,CommentProperty> map from parameter name to parameter
      */
     public function getMagicPropertyMap() : array
     {